# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API REST para la gestión de la clínica veterinaria: autenticación, usuarios, clientes, empleados, mascotas, citas, pagos y notificaciones.",
        "title": "API Clínica Veterinaria",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Equipo de Soporte API",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "2.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v2",
    "paths": {
...
        "/profile": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns the authenticated user's profile. Requires Bearer auth.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profile"
                ],
                "summary": "Get my profile",
                "responses": {
                    "200": {
                        "description": "Profile",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/clinic-vet-api_internal_shared_http.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/clinic-vet-api_internal_infrastructure_http_handlers_dtos.ProfileResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/clinic-vet-api_internal_shared_http.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/clinic-vet-api_internal_shared_http.APIResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Updates the authenticated user's profile (name, gender, date of birth, photo, bio). Requires Bearer auth.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profile"
                ],
                "summary": "Update my profile",
                "parameters": [
                    {
                        "description": "Profile fields to update",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/clinic-vet-api_internal_infrastructure_http_handlers_dtos.UpdateProfileRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Profile updated",
                        "schema": {
                            "$ref": "#/definitions/clinic-vet-api_internal_shared_http.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/clinic-vet-api_internal_shared_http.APIResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/clinic-vet-api_internal_shared_http.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/clinic-vet-api_internal_shared_http.APIResponse"
                        }
                    }
                }
            }
        },
        "/users": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Paginated search and filter of users. Requires admin or manager role.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Search users",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page_number",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Page size",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by email",
                        "name": "email",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by role",
                        "name": "role",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by status",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated list of users",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/clinic-vet-api_internal_shared_http.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/clinic-vet-api_internal_infrastructure_http_handlers_dtos.UserResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid query params",
                        "schema": {
                            "$ref": "#/definitions/clinic-vet-api_internal_shared_http.APIResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/clinic-vet-api_internal_shared_http.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/clinic-vet-api_internal_shared_http.APIResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a new user. Requires admin or manager role.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Create user",
                "parameters": [
                    {
                        "description": "User data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/clinic-vet-api_internal_infrastructure_http_handlers_dtos.CreateUserRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "User created",
                        "schema": {
                            "$ref": "#/definitions/clinic-vet-api_internal_shared_http.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/clinic-vet-api_internal_shared_http.APIResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/clinic-vet-api_internal_shared_http.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/clinic-vet-api_internal_shared_http.APIResponse"
                        }
                    }
                }
            }
        },
        "/users/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns a single user by ID. Requires admin or manager role.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Get user by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User found",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/clinic-vet-api_internal_shared_http.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/clinic-vet-api_internal_infrastructure_http_handlers_dtos.UserResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid ID",
                        "schema": {
                            "$ref": "#/definitions/clinic-vet-api_internal_shared_http.APIResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/clinic-vet-api_internal_shared_http.APIResponse"
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "schema": {
                            "$ref": "#/definitions/clinic-vet-api_internal_shared_http.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/clinic-vet-api_internal_shared_http.APIResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Soft-deletes a user by ID. Requires admin or manager role.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Delete user",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User deleted",
                        "schema": {
                            "$ref": "#/definitions/clinic-vet-api_internal_shared_http.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid ID",
                        "schema": {
                            "$ref": "#/definitions/clinic-vet-api_internal_shared_http.APIResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/clinic-vet-api_internal_shared_http.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/clinic-vet-api_internal_shared_http.APIResponse"
                        }
                    }
                }
            }
        },
        "/users/{id}/restore": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Restores a soft-deleted user. Requires admin or manager role.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Restore user",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User restored",
                        "schema": {
                            "$ref": "#/definitions/clinic-vet-api_internal_shared_http.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid ID",
                        "schema": {
                            "$ref": "#/definitions/clinic-vet-api_internal_shared_http.APIResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/clinic-vet-api_internal_shared_http.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/clinic-vet-api_internal_shared_http.APIResponse"
                        }
                    }
                }
            }
        },
        "/users/{id}/status": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Updates the status of a user (e.g. activate/deactivate). Requires admin or manager role.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Update user status",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "New status",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/clinic-vet-api_internal_infrastructure_http_handlers_dtos.UpdateUserStatusRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Status updated",
                        "schema": {
                            "$ref": "#/definitions/clinic-vet-api_internal_shared_http.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid ID or status",
                        "schema": {
                            "$ref": "#/definitions/clinic-vet-api_internal_shared_http.APIResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/clinic-vet-api_internal_shared_http.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/clinic-vet-api_internal_shared_http.APIResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "clinic-vet-api_internal_infrastructure_http_handlers_dtos.UpdateUserStatusRequest": {
            "description": "Request body for updating user account status",
            "type": "object",
            "required": [
                "status"
            ],
            "properties": {
                "status": {
                    "description": "New status (active, inactive, banned)\nRequired: true",
                    "type": "string",
                    "enum": [
                        "active",
                        "inactive",
                        "banned"
                    ],
                    "example": "active"
                }
            }
        },
...
        "clinic-vet-api_internal_shared_http.APIResponse": {
            "description": "Standardized API response structure.",
            "type": "object",
            "properties": {
                "data": {
                    "description": "The data payload of the response for successful requests."
                },
                "error": {
                    "description": "Details of the error if the request was not successful.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/clinic-vet-api_internal_shared_http.ErrorInfo"
                        }
                    ]
                },
                "message": {
                    "description": "A message providing additional context about the response.",
                    "type": "string"
                },
                "meta": {
                    "description": "Additional metadata for the response, such as pagination info."
                },
                "request_id": {
                    "description": "A unique identifier for the request.",
                    "type": "string"
                },
                "success": {
                    "description": "Indicates whether the request was successful.",
                    "type": "boolean"
                },
                "timestamp": {
                    "description": "The timestamp when the response was generated.",
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Bearer JWT token. Example: \"Bearer \u0026lt;access_token\u0026gt;\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}